+++ changelog.md
@@ -19,3 +19,7 @@
 - The constructor `TestFile(string filePath)` has been removed. Use `TestFile.Create(string filePath)` instead.
 - `TestFile`s can be created from an `Assembly` using `TestFile.Create(...)`.
 - Added `TestEnvironmentBuilder`.
+
+## Unreleased
+- Allow `MockModuleBuilder` to register anything `ContainerBuilder` can.
+- Add shortcuts to `ContainerBuilder` functionality e.g. `RegisterModule` and `RegisterInstance`.